cleanup: deleted unnecessary files and organized assets
